requirements.txt
pandas
psycopg2
sqlalchemy
openpyxl
numpy
xlrd
python-dotenv
requests
matplotlib
geopy
matplotlib
timezonefinder
fiona
shapely
seaborn
geopandas[all]
pymannkendall



